+++ app/src/main/kotlin/org/btcmap/Sync.kt
@@ -133,7 +133,7 @@ class Sync(val api: Api, val db: Database) {
                     db.comment.insert(newOrChanged.map {
                         Comment(
                             id = it.id,
-                            placeId = it.elementId,
+                            placeId = it.placeId,
                             comment = it.comment,
                             createdAt = ZonedDateTime.parse(it.createdAt),
                             updatedAt = ZonedDateTime.parse(it.updatedAt),