+++ app/src/main/kotlin/org/btcmap/comment/CommentsAdapterItem.kt
@@ -1,6 +1,7 @@
 package org.btcmap.comment
 
 data class CommentsAdapterItem(
+    val id: Long,
     val comment: String,
     val localizedDate: String,
 )